Add first version of userscript
This version executes a sequence of clicks that change the video quality as if a user interacted with the page and did it manually. It hides the settings menu while this takes place so the user experience is nearly seamless, experiencing only a video buffer as the feed is adjusted to a higher quality. To account for slower connections, the script waits for the page to load in before changing the quality.