Add ability to automatically target new versions of base.js
The server.js now uses a regex to detect where the setPlaybackQuality function is in the base.js and then finds what prototype g.k is set to. It then locates where the class referenced by g.k at that location is constructed and injects a snippet of code that exposes that class to the global scope.